Alternate Blind Proposal #1
All blinds and dazzles cast between players are always halved,
equivalent to a full-power Eagle Eyes at all times. Eagle Eyes still
resists blinds and dazzles cast by monsters, as well as adding offense
and damage to ranged attacks, but it does not stack with player innate
blind & dazzle resistance.

The purpose of this proposal is to specifically target the extreme
length of blind and dazzle durations against unbuffed targets, as well
as the fact that dazzle is never reduced by Eagle Eyes because a dazzler
can also purge. With this proposal, it's as if all players have Eagle
Eyes all the time.

This proposal:
- avoids any nerfs to beloved spells
- uses with the same durations we've all gotten used to
- protects builders and unaware players
- relieves some of the excessive power of purge + dazzle