Alternate Blind Proposal #4
Enable armor spell procs.

What these do is engrave a random spell on your armor with a random
spellpower and type. Some combinations are useless. Some are awesome.
The concern is with unbuffed targets being Blinded for so long, and with
Purge; well, these armors will help rebuff you during combat when you
are struck.

The random nature creates a natural pursuit of good procs. You won't
find a '99 eagle eyes 15% of the time on yourself' all the time.
Builders will find great armors naturally, but others may have to buy or
farm theirs (NPC-sold armor CAN have procs).

-------

Examples:

"When defending against a strike, the runes engraved on this ivy circlet
will cast eagle eyes on you with 73 spellpower 15% of the time."

"When defending against a strike, the runes engraved on this scale armor
will cast resist cold on you with 28 spellpower 6% of the time."

"When defending against a strike, the runes engraved on this magic
spirit helmet will cast dispel illusion with 49 spellpower 1% of the
time."

"When defending against a strike, the runes engraved on this small round
shield will cast blink with 52 spellpower 20% of the time."

"When defending against a strike, the runes engraved on this troll mask
will cast eagle eyes on the enemy with 97 spellpower 7% of the time."
fitting for a troll mask ;)

"When defending against a strike, the runes engraved on this pants will
cast remove curse on you with 66 spellpower 2% of the time."

"When defending against a strike, the runes engraved on this light
jerkin will cast super strength on you with 84 spellpower 4% of the
time."

------

Each armor class has its own set of spells and chances.

Masks can be worn with circlets now, remember, so when we put masks in,
they'll have a use - they'll have an extra proc :)

This proposal:
- nerfs no spells directly
- creates a new treasure and loot system
- opens up combat to a huge variety of comebacks
- adds lots of utility (remove curse pants, spiritual hammer shirts, and
what not)
- basically it's really cool