Deflect Rebalancing
Deflect is still Kraanan level 4 and still costs 20 mana, but it is no
longer a Personal Enchantment. Instead, it is a Radius Enchantment with
a static range of 6 (splash spells are range 4).

Deflect now creates a bubble that attempts to reflect all single-target
attack spells of any type that are cast from outside that bubble. If you
Lightning Bolt or Blast of Fire or Splash of Acid inside the bubble,
Deflect does nothing. You can't cast splash spells from greater than
range 6, so they are unaffected generally, but Deflect will attempt to
reflect Lightning Bolts and Fireballs that CAN be cast from that far
away.

Your Deflect bubble will protect all of your allies and your enemies.

Casting Deflect and affecting an innocent can cause you to be labeled an
outlaw, to prevent muling issues. Keep your safety on.