Shatterlock Restructured as Event Spell
Shatterlock opening Guildhalls was a bad idea for many reasons, and
there is no feasible implementation that will preserve its use without
being a broken mechanic.

So, instead, Shatterlock can now be used to initiate events in specific
rooms. The first instance of this will be the ability to open the
Martyr's Battleground node by shatterlocking the magic of the stone
henge in Ukgoth.

Prisms also no longer reset on the ground, and the trance for prism
spells has been significantly upgraded - you can now only lose trance by
resting or running out of mana, so combat is still possible while
focusing.