Spell Item Splitter owner checks
If the player is holding a spell item, it will split into his inventory.

If the spell item is on the floor in a room, it will split onto the
floor.

If it's not owned by a player or by a room, the new potion will delete
rather than go anywhere.