Spellbane Improvement
Spellbane is now a very unique song in several ways.

* Its range starts at 2 and increases to 3 (mystic range, but just short
of splash range) at 99%. This radius is based only on % in the song, not
spellpower.

* Spellbane lowers the power of any sung cast under its influence. The
new power of spells becomes equal to Spellbane's own power. In other
words, you want your Spellbane to be as weak as possible. Pile on AMA
and Plate to reach 0 spellpower, and then all spells cast in your
vicinity will also have 0 spellpower (including your own).

* Spellbane can only lower spellpower, never raise it.

* Note that 0 spellpower does NOT actually prevent casting entirely like
old Spellbane (that was too boring). Spells can still be cast, albeit
painfully. For example, a blast of fire that does not fizzle will do
only half its normal damage at 0 spellpower. Be wary of spells with high
minimum effects, or spells that can't fizzle.

So what is this intended to do? Well, first off, this is part of my
(hopeful) new push into upgrading Jala into something many characters
consider taking. This is the first example of a spell that benefits
fighters, something they might take as /J3 as an option for their build.
At the same time, it's useless to pure mages, which is important, since
they're already Jala 6 for MC and CM.