Reincarnation system proposal
* Players may qualify for reincarnation by going to any skill or spell
teacher while at max base health and maxed schools. The teacher will
flag the player for reincarnation after assessing that he or she can no
longer learn abilities. After that, the five major priestesses (and
Parrin) can reincarnate the player who states "I wish to be
reincarnated." (note, the flag is not lost if the player dies or pens
after qualifying).

* A reincarnated player is recalibrated, deliveranced, set to 20 health
and zero spells and zero skills, and then given 1 stat point based on
the priestess employed. Guild affiliation and items are retained (this
is not a suicide). Players who reincarnate can never again be angeled.
You will be attackable at 20 hp.

* Players can reincarnate up to ten times, for ten stat points total.

* Player bios state "This player has reincarnated X times."