Monster Radius Spells
All of the spells are designed so that players can 'see them' to discord
them, but they offer benefits only to monster allies. Monster allies
includes you in a rare case - if you have charmed a monster and are its
master.

* Orc Wizards can now cast Primal Restoration, which heals all orc
allies in the vicinity at a rapid rate.

* Avar Shamans can now cast Blur, which boosts the defense of all
members of its clan in the vicinity by 300 (75 spellpower * 4).

* Xeo'chicatl can now cast resistance spheres which protect all
Xeochicatls in range from a specific element by a large degree (+99%).
Each Xeo'chicatl casts a resistance sphere of the element it is weak to.
For example, Ma'Xeochicatls are weak to Fire, and thus they cast a fire
resistance sphere. Their -20 weakness + 99 = 79%, so they themselves are
still weak to that element by a certain amount.

I also changed Sky xeos to be weak to acid instead of nerudite to
complete the analogy to the four major elements. So:

Earth xeos are weak to shock and cast shock resistance sphere
Water xeos are weak to fire and cast fire resistance sphere
Fire xeos are weak to cold and cast cold resistance sphere
Sky xeos are weak to acid and cast acid resistance sphere

You will of course have a major problem on your hands if you're a mage
and fighting a large group of active xeos and the spheres start to
overlap. You're going to want to move and fight carefully, or discord.