Silence safety issues fixed
Silence has been converted to RE with a 1 mana per second drain. It no
longer breaks Jala songs and prevents all casting. For the time being,
it only prevents communication in a large range around the caster. We
can't have a level 3 spell like this nuking all magic... if someone
thinks of a good use for it, that can be added later.

For now, it still remains unique in that it silences communications.
Perhaps enterprising fighters will use it to prevent calls for help from
victims.