Illusionary Form changes
* Dusk rats can no longer be morphed or illusionary formed.

* There is now a new type of enemy called a corrupted avar chieftain.
It's a stronger avar chieftain with a poison similar to the dusk rat's.
It can be morphed into with the word teotkoavar. The corrupted avar
chieftain cannot multi-poison a player (only one poison at a time).
These will spawn at Ruins of the Riija temple and other avars will fight
them.