Player Housing Prototype
* Buy a house buy saying "buy house" in Familiars. Costs 500k.

* Given a key that, when used in Familiars, takes you to your house.

* Buy extra rooms by saying "buy room" inside. Costs 1 million for the
first expansion. The second expansion adds a new room on a second floor,
and costs 2 million. The final expansion adds a second room on the
second floor.

* Rooms 3 and 4 are on the second floor of your house - go up the stairs
to reach them. This is the first application of a new modular technique
that actually gives us the illusion of a functioning Z axis. Your second
floor is directly above your first floor - it just changes heights and
loads whichever items are on that floor as you use the stairs.