Flame Dash spell
This is an example of a few new techniques.

1: Attempting to build in-map teleportation. Flame Dash attempts to leap
you forward up to 5 units, based on spellpower. Unfortunately,
CanMoveInRoomHighRes doesn't seem to be giving proper checks for players
on this.

2: This is an example of a multi-cooldown system, where you can dash
multiple times, but then have to wait if you blow them all too quickly.

3: Flame dash leaves short-lived behind firewall elements in the path
you travel. I'd like to experiment with more mechanics like these, that
actually change the battlefield temporarily.