Rod rebalance
Rods were given a delay, which is probably a good thing. Here's a shot
at improving what we're intended to do with them.

* Set base rod delay to 500ms. (As you'll see below, this really isn't
an important number anymore).

* Added base rod delay into effect durations, so players will never be
penalized for rod delays, now or in the future.

* Heal rods can store only 1 total use (7 charges required). They now
heal five times as much (20-40) in that single use, instead of healing
4-8 five times. In essence, this removes the "spammy" nature of wands
once and for all without devaluing the health they give. They were also
trivializing building a bit, so their kills required were upped from 5
to 7 (40% more kills required). If builders are carrying 5, they will
still be able to use a 20-40 heal for 5 of the 7 mobs they're fighting.
Getting a 20-40 heal for every single mob might have been too much.

* Shield Rod given a base delay of 2000 ms. Because base delay is added
into effect duration, this extends shield rods up to 7 seconds. However,
the cost of that is an inability to attack or cast for the first 2
seconds. This is a pseudo-buff with the intention of giving the shield a
physical "feel" (i.e. no slamming shield rod and continuing to fight
immediately).

* Vamp rod given 50 ms delay. These don't do anything particularly
combat related.