Might proposal
We've thought about this for a LONG time. Personally, as a builder, I
just won't bother with Might as long as it's a tiny bit of damage hiding
somewhere in base weapon attacks.

This pull adds Might as a FINAL damage modifer for melee weapons, up to
and exactly equal to a Rod of Power (+35% at 70 Might). They do not
overlap. You either get up to 35% from Might, your get the full 35% from
Rod of Power. Ranged and spells continue to function the same with Rod
of Power.

This also leaves currently leaves in base stat effects. Agility does
still affect scimitar and axe, might affects hammer and mace, aim
affects mystic sword. Then might affects ALL of them with the final
damage bonus (so might is double counted for hammer/mace, yes). This is
purely a buff to existing weapons at this time.

That may change during testing.

The idea here is to make melee weapons terrifying in the hands of
someone with very high stats. 70 might + 50 agility better hit hard with
a scimitar, that's half of their available points alone.