Spellbane description fix, range calculation change, new mana cost
Much discussion has been had over Spellbane, and I tried to combine the
best ideas and concerns I heard. I'll try to keep this short.

Let's start with asking who will be using this spell. The general
continuum includes pure mages, hybrids, or pure fighters.

Pure mages will never, ever use this spell in real combat. That leaves
mule situations. For this, we add an additional mana cost component
equal to the casting cost of the blocked spell, so that groups can
'power through' even a mule's spellbane by spamming high cost spells.
Remember that groups can also still strike with weapons, and Spellbane
and Truce cannot be in effect at the same time (Spellbane is hostile).
We also change the range calculation so that using Spellbane in
conjunction with AMA is exceedingly difficult.

Hybrid characters might actually employ this spell in conjunction with
bow or melee attacks. They have the mana to support it (even with the
additional mana cost) and they have weapon attacks to continue fighting.
However, hybrids probably need their spells, so use of Spellbane by them
will be situational, and probably mostly in the open wilds where blind
and hold are less necessary. It's also rare that a hybrid has Jala 5.

Pure fighters actually have the most interest in this spell, which pairs
well (from our standpoint) with their typically low mana. Jala Pure
Fighters do exist, although Jala 4 and 5 are more or less wasted levels
for them. Spellbane should probably be lowered to level 3 to make this a
more easily takeable 'package' for Pure Fighters (Jala 1-3 is actually
pretty useful for them). The new mana cost of the song synergizes with
Jala 6, so Jala 6 PFs (which are already a thing in some ways) can
actually use Spellbane. Fortunately, our range change to base the song
on spellpower makes it incompatible with AMA. This is an 'alternate
strategy' at fighting against magic. Spellcasting enemies can get far
enough away or simply power through your small mana reserves, yes, but
Spellbane can buy you time and make melee range punishing for your
opponent. The other great thing about Spellbane is the tension it has
with many strategies: if you're Spellbaning, you can't Discord or AMA.
Use of this spell is not an automatic gimme.

I don't like blanket anti-magic abilities in general, but I think this
might be a good point to start with for Spellbane. I recommend
incremental upgrades or downgrades based on player feedback in the near
future. I would not consider how often this song is used a viable
metric, though, as things like these are very high level in terms of
meta and strategy (or they should be, anyway). The game benefits from
having extremely powerful options that take skill and planning to use.